requirements.txt
Flask==2.0.2
flask-redis==0.4.0
python-dotenv==0.21.0
